requirements.txt
flask
python-dotenv
fastapi
jsonify
uvicorn
prometheus_client